`train()` changed with the algorithm

+++ src/lib/utils.rs
@@ -113,10 +113,10 @@ pub(crate) fn translate<L: Literal>(vec: Vec<L>) -> Vec<Vec<u32>> {
         return result;
 }
 
-pub(crate) fn sum(vec: &Vec<u32>) -> u32 {
+pub(crate) fn sum(vec: &Vec<u32>) -> f32 {
     let mut sum: u32 = 0;
     for i in vec {
         sum += i;
     };
-    return sum;
+    return sum as f32;
 }
